change to mace version

+++ ParTner/openCVLibrary341/java/src/org/opencv/core/MatOfRect2d.java
@@ -49,10 +49,10 @@ public class MatOfRect2d extends Mat {
         double buff[] = new double[num * _channels];
         for(int i=0; i<num; i++) {
             Rect2d r = a[i];
-            buff[_channels*i+0] = r.x;
-            buff[_channels*i+1] = r.y;
-            buff[_channels*i+2] = r.width;
-            buff[_channels*i+3] = r.height;
+            buff[_channels*i+0] = (double) r.x;
+            buff[_channels*i+1] = (double) r.y;
+            buff[_channels*i+2] = (double) r.width;
+            buff[_channels*i+3] = (double) r.height;
         }
         put(0, 0, buff); //TODO: check ret val!
     }
